added eclipse .project file added account transfer functionality refactored amounts to use BigDecimal instead of double in preparation for calculating accrued interest

+++ src/test/java/com/abc/BankTest.java
@@ -4,6 +4,8 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
+import java.math.BigDecimal;
+
 public class BankTest {
     private static final double DOUBLE_DELTA = 1e-15;
 
@@ -24,9 +26,9 @@ public class BankTest {
         Customer bill = new Customer("Bill").openAccount(checkingAccount);
         bank.addCustomer(bill);
 
-        checkingAccount.deposit(100.0);
+        checkingAccount.deposit(BigDecimal.valueOf(100));
 
-        assertEquals(0.1, bank.totalInterestPaid(), DOUBLE_DELTA);
+        assertEquals(0.1, bank.totalInterestPaid().doubleValue(), DOUBLE_DELTA);
     }
 
     @Test
@@ -35,9 +37,9 @@ public class BankTest {
         Account checkingAccount = new Account(Account.SAVINGS);
         bank.addCustomer(new Customer("Bill").openAccount(checkingAccount));
 
-        checkingAccount.deposit(1500.0);
+        checkingAccount.deposit(BigDecimal.valueOf(1500));
 
-        assertEquals(2.0, bank.totalInterestPaid(), DOUBLE_DELTA);
+        assertEquals(2.0, bank.totalInterestPaid().doubleValue(), DOUBLE_DELTA);
     }
 
     @Test
@@ -46,9 +48,9 @@ public class BankTest {
         Account checkingAccount = new Account(Account.MAXI_SAVINGS);
         bank.addCustomer(new Customer("Bill").openAccount(checkingAccount));
 
-        checkingAccount.deposit(3000.0);
+        checkingAccount.deposit(BigDecimal.valueOf(3000));
 
-        assertEquals(170.0, bank.totalInterestPaid(), DOUBLE_DELTA);
+        assertEquals(170.0, bank.totalInterestPaid().doubleValue(), DOUBLE_DELTA);
     }
 
 }
